Refactor vmware_cluster into several modules
